Rework user stats overview page
