+++ src/pages/User.jsx
@@ -1,10 +1,10 @@
 import { A, useLocation, useNavigate, useParams, useSearchParams } from "@solidjs/router";
 import api, { IndexedDB } from "../utils/api.js";
-import { createContext, createEffect, createMemo, createSignal, For, Match, on, onCleanup, Show, untrack, useContext } from "solid-js";
+import { createContext, createEffect, createSignal, For, Match, on, onCleanup, Show, untrack, useContext } from "solid-js";
 import "./User.scss";
 import { useAuthentication } from "../context/AuthenticationContext.jsx";
 import { assert } from "../utils/assert.js";
-import { capitalize, countryNameFromCountryCode, formatMediaFormat, formatTimeToDate, formatTitleToUrl, numberCommas } from "../utils/formating.js";
+import { formatTimeToDate, formatTitleToUrl, numberCommas } from "../utils/formating.js";
 import { ActivityCard } from "../components/Activity.jsx";
 import UserMediaListWorker from "../worker/user-media-list.js?worker";
 import { useEditMediaEntries } from "../context/EditMediaEntriesContext.jsx";
@@ -42,7 +42,7 @@ export function User(props) {
   )
 }
 
-const useUser = () => useContext(UserContext);
+export const useUser = () => useContext(UserContext);
 
 function Content(props) {
   const { user, following } = useUser();
@@ -1083,399 +1083,5 @@ export function Stats(props) {
         {props.children}
       </div>
     </div>
-
-  );
-}
-
-export function StatsOverview() {
-  const params = useParams();
-  const { user } = useUser();
-  const { accessToken } = useAuthentication();
-  const [userStats] = api.anilist.userAnimeStats(() => params.name, accessToken);
-
-  return (
-    <Show when={userStats()}>
-      <section class="user-profile-stats-header-section">
-        <ul class="user-profile-stats-general-header">
-          <li>
-            <div class="svg-container">
-              <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 640 512"><path fill="currentColor" d="M592 0H48C21.5 0 0 21.5 0 48v320c0 26.5 21.5 48 48 48h245.1v32h-160c-17.7 0-32 14.3-32 32s14.3 32 32 32h384c17.7 0 32-14.3 32-32s-14.3-32-32-32h-160v-32H592c26.5 0 48-21.5 48-48V48c0-26.5-21.5-48-48-48zm-16 352H64V64h512v288z"></path></svg>
-            </div>
-            <div class="right">
-              <p class="value">{numberCommas(user().statistics.anime.count || 0)}</p>
-              <p class="title">Total anime</p>
-            </div>
-          </li>
-          <li>
-            <div class="svg-container">
-              <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M424.4 214.7L72.4 6.6C43.8-10.3 0 6.1 0 47.9V464c0 37.5 40.7 60.1 72.4 41.3l352-208c31.4-18.5 31.5-64.1 0-82.6z"></path></svg>
-            </div>
-            <div class="right">
-              <p class="value">{numberCommas(user().statistics.anime.episodesWatched || 0)}</p>
-              <p class="title">Episodes watched</p>
-            </div>
-          </li>
-          <li>
-            <div class="svg-container">
-              <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M12 192h424c6.6 0 12 5.4 12 12v260c0 26.5-21.5 48-48 48H48c-26.5 0-48-21.5-48-48V204c0-6.6 5.4-12 12-12zm436-44v-36c0-26.5-21.5-48-48-48h-48V12c0-6.6-5.4-12-12-12h-40c-6.6 0-12 5.4-12 12v52H160V12c0-6.6-5.4-12-12-12h-40c-6.6 0-12 5.4-12 12v52H48C21.5 64 0 85.5 0 112v36c0 6.6 5.4 12 12 12h424c6.6 0 12-5.4 12-12z"></path></svg>
-            </div>
-            <div class="right">
-              <p class="value">{((user().statistics.anime.minutesWatched || 0) / 60 / 24).toFixed(1)}</p>
-              <p class="title">Days watched</p>
-            </div>
-          </li>
-          <li>
-            <div class="svg-container">
-              <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 384 512"><path fill="currentColor" d="M360 64c13.255 0 24-10.745 24-24V24c0-13.255-10.745-24-24-24H24C10.745 0 0 10.745 0 24v16c0 13.255 10.745 24 24 24 0 90.965 51.016 167.734 120.842 192C75.016 280.266 24 357.035 24 448c-13.255 0-24 10.745-24 24v16c0 13.255 10.745 24 24 24h336c13.255 0 24-10.745 24-24v-16c0-13.255-10.745-24-24-24 0-90.965-51.016-167.734-120.842-192C308.984 231.734 360 154.965 360 64z"></path></svg>
-            </div>
-            <div class="right">
-              <p class="value">{((userStats().data.statuses.find(list => list.status === "PLANNING")?.minutesWatched || 0) / 60 / 24).toFixed(1)}</p>
-              <p class="title">Days planned</p>
-            </div>
-          </li>
-          <li>
-            <div class="svg-container">
-              <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 384 512"><path fill="currentColor" d="M109.25 173.25c24.99-24.99 24.99-65.52 0-90.51-24.99-24.99-65.52-24.99-90.51 0-24.99 24.99-24.99 65.52 0 90.51 25 25 65.52 25 90.51 0zm256 165.49c-24.99-24.99-65.52-24.99-90.51 0-24.99 24.99-24.99 65.52 0 90.51 24.99 24.99 65.52 24.99 90.51 0 25-24.99 25-65.51 0-90.51zm-1.94-231.43l-22.62-22.62c-12.5-12.5-32.76-12.5-45.25 0L20.69 359.44c-12.5 12.5-12.5 32.76 0 45.25l22.62 22.62c12.5 12.5 32.76 12.5 45.25 0l274.75-274.75c12.5-12.49 12.5-32.75 0-45.25z"></path></svg>
-            </div>
-            <div class="right">
-              <p class="value">{(user().statistics.anime.meanScore || 0).toFixed(2)}</p>
-              <p class="title">Mean score</p>
-            </div>
-          </li>
-          <li>
-            <div class="svg-container">
-              <svg aria-hidden="true" focusable="false" role="img" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path fill="currentColor" d="M224 352c-35.35 0-64 28.65-64 64s28.65 64 64 64 64-28.65 64-64-28.65-64-64-64zm0-192c35.35 0 64-28.65 64-64s-28.65-64-64-64-64 28.65-64 64 28.65 64 64 64zm192 48H32c-17.67 0-32 14.33-32 32v32c0 17.67 14.33 32 32 32h384c17.67 0 32-14.33 32-32v-32c0-17.67-14.33-32-32-32z"></path></svg>
-            </div>
-            <div class="right">
-              <p class="value">{(user().statistics.anime.standardDeviation || 0).toFixed(1)}</p>
-              <p class="title">Standard Deviation</p>
-            </div>
-          </li>
-        </ul>
-      </section>
-      <StatsScoreDistributionBars data={userStats().data.scores.sort((a, b) => a.score - b.score)}/>
-      <StatsEpisodeCountBars data={userStats().data.lengths.sort((a, b) => (parseInt(a.length) || Infinity) - (parseInt(b.length) || Infinity))}/>
-      <StatsDistributionLists formats={userStats().data.formats} statuses={userStats().data.statuses} countries={userStats().data.countries} />
-      <StatsYearLineCharts heading="Release year" data={userStats().data.releaseYears.sort((a, b) => a.releaseYear - b.releaseYear)}/>
-      <StatsYearLineCharts heading="Watch year" data={userStats().data.startYears.sort((a, b) => a.startYear - b.startYear)}/>
-    </Show>
-  );
-}
-
-function useWidth(elem) {
-  const [width, setWidth] = createSignal(elem()?.getBoundingClientRect().width || 0);
-  const resize = () => {
-    setWidth(elem()?.getBoundingClientRect().width || 0);
-  }
-
-  createEffect(on(elem, resize));
-
-  window.addEventListener("resize", resize);
-
-  onCleanup(() => {
-    window.removeEventListener("resize", resize);
-  });
-
-  return width;
-}
-
-const lerp = (a, b, t) => {
-  return a + t * (b - a);
-}
-
-function StatsScoreDistributionBars(props) {
-  const [state, setState] = createSignal("count");
-  const [max, setMax] = createSignal(0);
-
-  createEffect(() => {
-    const maxValue = props.data.reduce((acc, v) => Math.max(acc, v[state()]), 0);
-    setMax(maxValue);
-  });
-
-  return (
-    <section class="user-profile-stat-score-bar-section">
-      <div class="header flex-space-between">
-        <h2>Score distributions</h2>
-        <div>
-          <button onClick={() => setState("count")}>Titles Watched</button>
-          <button onClick={() => setState("minutesWatched")}>Hours Watched</button>
-        </div>
-      </div>
-      <DraggableScrollContainer>
-        <ol>
-          <For each={props.data}>{stat => (
-            <li>
-              <p>{stat.score}</p>
-              <div style={{height: `${stat[state()] / max() * 85}%`}}></div>
-              <p>
-                <Show when={state() === "minutesWatched"} fallback={stat[state()]}>
-                  {numberCommas(Math.ceil(stat[state()] / 60))}
-                </Show>
-              </p>
-            </li>
-          )}</For>
-        </ol>
-      </DraggableScrollContainer>
-    </section>
-  );
-}
-
-function StatsEpisodeCountBars(props) {
-  const [state, setState] = createSignal("count");
-  const [max, setMax] = createSignal(0);
-
-  createEffect(() => {
-    const maxValue = props.data.reduce((acc, v) => Math.max(acc, v[state()]), 0);
-    setMax(maxValue);
-  });
-
-  return (
-    <section class="user-profile-stat-score-bar-section">
-      <div class="header flex-space-between">
-        <h2>Episode count</h2>
-        <div>
-          <button onClick={() => setState("count")}>Titles Watched</button>
-          <button onClick={() => setState("minutesWatched")}>Hours Watched</button>
-          <button onClick={() => setState("meanScore")}>Mean Score</button>
-        </div>
-      </div>
-      <DraggableScrollContainer>
-        <ol>
-          <For each={props.data}>{stat => (
-            <li>
-              <p>{stat.length || "Unknown"}</p>
-              <div style={{height: `${stat[state()] / max() * 85}%`}}></div>
-              <p>
-                <Show when={state() === "minutesWatched"} fallback={stat[state()]}>
-                  {numberCommas(Math.ceil(stat[state()] / 60))}
-                </Show>
-              </p>
-            </li>
-          )}</For>
-        </ol>
-      </DraggableScrollContainer>
-    </section>
-  );
-}
-
-function StatsYearLineCharts(props) {
-  let container;
-  const [max, setMax] = createSignal(0);
-  const containerWidth = useWidth(() => container);
-  const [state, setState] = createSignal("count");
-
-  const inlinePadding = 32;
-  const topPadding = 64;
-  const bottomPadding = 60;
-  const width = () => Math.max(50, containerWidth() / props.data.length);
-  const getX = (x) => inlinePadding + x * width();
-  const getY = (stat) => Math.ceil((1 - stat / max()) * 200 + topPadding);
-
-  createEffect(() => {
-    const maxValue = props.data.reduce((acc, v) => Math.max(acc, v[state()]), 0);
-    setMax(maxValue);
-  });
-
-  createEffect(on(containerWidth, () => {
-    container?.classList.add("no-motion");
-
-    setTimeout(() => {
-      container?.classList.remove("no-motion");
-    }, 100);
-  }));
-
-  const path = createMemo(() => {
-    const rounding = .35;
-    return props.data.map((year, i, arr) => {
-      if (i === 0) {
-        return "M" + getX(i) + " " + getY(year[state()]);
-      } if (i < arr.length - 1) {
-        return "S" + lerp(getX(i), getX(i - 1), rounding) + " " + lerp(getY(year[state()]), getY(year[state()]) + (getY(arr[i - 1][state()]) - getY(arr[i + 1][state()])) / 2, rounding) + "," + getX(i) + " " + getY(year[state()]);
-      }
-      return "S" + lerp(getX(i), getX(i - 1), rounding) + " " + lerp(getY(year[state()]), getY(arr[i - 1][state()]), rounding) + "," + getX(i) + " " + getY(year[state()]);
-    }).join("");
-  });
-
-  const pathFill = createMemo(() => {
-    return path() + "L" + getX(props.data.length - 1) + " " + getY(0) + bottomPadding + "L" + inlinePadding + " " + getY(0) + bottomPadding;
-  });
-
-  return (
-    <Show when={props.data.length}>
-      <section class="user-profile-stats-graph-container no-motion" ref={container}>
-        <div class="header flex-space-between">
-          <h2>{props.heading}</h2>
-          <div>
-            <button onClick={() => setState("count")}>Titles Watched</button>
-            <button onClick={() => setState("minutesWatched")}>Hours Watched</button>
-            <button onClick={() => setState("meanScore")}>Mean Score</button>
-          </div>
-        </div>
-        <DraggableScrollContainer>
-          <svg width={getX(props.data.length - 1) + inlinePadding} height={getY(0) + bottomPadding}>
-            <path d={pathFill()} stroke="none" stroke-width="0" fill="grey" />
-            <rect x="0" y={getY(0)} width="100%" height="60" fill="darkgrey" stroke="none" pointer-events="all" />
-            <path d={path()} stroke="black" stroke-width="5" fill="transparent" />
-            <For each={props.data}>{(year, i) => (
-              <g class="item">
-                <rect x={getX(i()) - width() / 2} y="0" width={width()} height="100%" fill="none" stroke="none" pointer-events="all" />
-                <circle cx={getX(i())} cy={getY(year[state()])} r="6" pointer-events="none" />
-                <text class="text" x={getX(i())} y="0" style={{translate: `0 ${getY(year[state()]) - 10}px`}} text-anchor="middle">
-                  <Show when={state() === "minutesWatched"} fallback={year[state()]}>
-                    {numberCommas(Math.ceil(year[state()] / 60))}
-                  </Show>
-                </text>
-                <text class="year" x={getX(i())} y="304" text-anchor="middle">{year.releaseYear || year.startYear}</text>
-              </g>
-            )}</For>
-          </svg>
-        </DraggableScrollContainer>
-      </section>
-    </Show>
-  );
-}
-
-function DraggableScrollContainer(props) {
-  let xStart = 0;
-  let xScroll = 0;
-  let xPrev = NaN;
-  let scrollContainer;
-
-  return (
-    <div class="scroll" ref={scrollContainer} onMouseMove={e => {
-      if (e.buttons === 1) {
-        e.preventDefault();
-        const xDelta = e.clientX - xStart;
-        xPrev = e.clientX;
-        scrollContainer.style.userSelect = "none";
-        scrollContainer.scrollTo(xScroll - xDelta, 0);
-      } else {
-        scrollContainer.style.userSelect = null;
-        xStart = e.clientX;
-        xScroll = scrollContainer.scrollLeft;
-
-        const xDelta = e.clientX - xPrev;
-        xPrev = NaN;
-        if (Math.abs(xDelta) > .1) {
-          const momentum = (start, time, deltaMomentum) => {
-            if (Math.abs(deltaMomentum) < 0.5) {
-              return;
-            } 
-            scrollContainer.scrollBy(-deltaMomentum * 2, 0);
-            requestAnimationFrame((t) => momentum(start, t, (deltaMomentum * (time - start < 200 ? 0.99 : 0.95))));
-          }
-          requestAnimationFrame((t) => momentum(t, t, xDelta));
-        }
-      }
-    }}>
-      {props.children}
-    </div>
-  );
-}
-
-const plural = num => num !== 1 ? "s" : "";
-
-function StatsDistributionLists(props) {
-  const [formats, setFormats] = createSignal();
-  const params = useParams();
-  const { user } = useUser();
-
-  createEffect(() => {
-    setFormats(props.formats.reduce((acc, v) => acc + v.count, 0));
-  });
-
-  return (
-    <section class="user-profile-stats-formats">
-      <div>
-        <h2>Format distribution</h2>
-        <ol>
-          <For each={props.formats}>{format => (
-            <li>
-              <div>
-                <div class="container">
-                  <A class="title" href={"/user/" + user().name + "/anime?format=" + format.format}>{formatMediaFormat(format.format)}</A>
-                  <p>{format.meanScore || ""}</p>
-                </div>
-                <p class="time">
-                  <Show when={Math.floor(format.minutesWatched / 60 / 24)}>{days => <>{numberCommas(days())} day{plural(days())} </>}</Show>
-                  <Show when={Math.floor(format.minutesWatched / 60 % 24)}>{hours => <>{numberCommas(hours())} hour{plural(hours())} </>}</Show>
-                  <Show when={format.minutesWatched < 60}>{format.minutesWatched} minute{plural(format.minutesWatched)}</Show>
-                </p>
-              </div>
-              <div class="right">
-                <p>{(format.count / formats() * 100).toFixed(2)}%</p>
-                <p>{numberCommas(format.count)}/{formats()}</p>
-              </div>
-            </li>
-          )}</For>
-        </ol>
-        <div class="filler"></div>
-      </div>
-      <div>
-        <h2>Status distribution</h2>
-        <ol>
-          <For each={props.statuses}>{status => (
-            <li>
-              <div>
-                <div class="container">
-                  <A class="title" href={"/user/" + user().name + "/anime?userStatus=" + status.status}>
-                    <Switch fallback={capitalize(status.status)}>
-                      <Match when={status.status === "CURRENT"}>
-                        <Switch>
-                          <Match when={params.type === "anime"}>Watching</Match>
-                          <Match when={params.type === "manga"}>Reading</Match>
-                        </Switch>
-                      </Match>
-                      <Match when={status.status === "REPEATING"}>
-                        <Switch>
-                          <Match when={params.type === "anime"}>Rewatching</Match>
-                          <Match when={params.type === "manga"}>Rereading</Match>
-                        </Switch>
-                      </Match>
-                    </Switch>
-                  </A>
-                  <p>{status.meanScore || ""}</p>
-                </div>
-                <p class="time">
-                  <Show when={Math.floor(status.minutesWatched / 60 / 24)}>{days => <>{numberCommas(days())} day{plural(days())} </>}</Show>
-                  <Show when={Math.floor(status.minutesWatched / 60 % 24)}>{hours => <>{numberCommas(hours())} hour{plural(hours())} </>}</Show>
-                  <Show when={status.minutesWatched < 60}>{status.minutesWatched} minute{plural(status.minutesWatched)}</Show>
-                </p>
-              </div>
-              <div class="right">
-                <p>{(status.count / formats() * 100).toFixed(2)}%</p>
-                <p>{numberCommas(status.count)}/{formats()}</p>
-              </div>
-            </li>
-          )}</For>
-        </ol>
-        <div class="filler"></div>
-      </div>
-      <div>
-        <h2>Country distribution</h2>
-        <ol>
-          <For each={props.countries}>{country => (
-            <li>
-              <div>
-                <div class="container">
-                  <A class="title" href={"/user/" + user().name + "/anime?countryOfOrigin=" + country.country}>{countryNameFromCountryCode(country.country)}</A>
-                  <p>{country.meanScore || ""}</p>
-                </div>
-                <p class="time">
-                  <Show when={Math.floor(country.minutesWatched / 60 / 24)}>{days => <>{numberCommas(days())} day{plural(days())} </>}</Show>
-                  <Show when={Math.floor(country.minutesWatched / 60 % 24)}>{hours => <>{numberCommas(hours())} hour{plural(hours())} </>}</Show>
-                  <Show when={country.minutesWatched < 60}>{country.minutesWatched} minute{plural(country.minutesWatched)}</Show>
-                </p>
-              </div>
-              <div class="right">
-                <p>{(country.count / formats() * 100).toFixed(2)}%</p>
-                <p>{numberCommas(country.count)}/{formats()}</p>
-              </div>
-            </li>
-          )}</For>
-        </ol>
-        <div class="filler"></div>
-      </div>
-    </section>
   );
 }
